requirements.txt
amqp==5.3.1
asgiref==3.6.0
billiard==4.2.1
celery==5.4.0
cffi==1.15.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==40.0.2
Django==4.2.1
django-appconf==1.0.5
django-ckeditor==6.5.1
django-froala-editor==4.0.19
django-js-asset==2.0.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
kombu==5.4.2
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
prompt_toolkit==3.0.50
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
redis==5.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.4.0
